+++ lib/decorators.py
@@ -1,0 +1,24 @@
+import time
+
+
+class classproperty(property):
+
+    def __get__(self, cls, owner):
+        return self.fget.__get__(None, owner)()
+
+
+def print_time(func):
+     """
+     @print_time
+
+     Put this decorator around a function to see how many seconds each
+     call of this function takes to run.
+     """
+     def wrapped_func(*args, **kwargs):
+         start = time.time()
+         result = func(*args, **kwargs)
+         end = time.time()
+         seconds = end - start
+         print "SECONDS:", seconds, func.__name__, kwargs
+         return result
+     return wrapped_func
