+++ lib/utils.py
@@ -1,7 +1,6 @@
 import json
 import hashlib
 from math import isnan
-import time
 
 import numpy as np
 from bson import json_util
@@ -76,27 +75,3 @@ def series_to_jsondict(s):
 
 def df_to_jsondict(df):
     return [series_to_jsondict(s) for i, s in df.iterrows()]
-
-
-
-class classproperty(property):
-
-    def __get__(self, cls, owner):
-        return self.fget.__get__(None, owner)()
-
-
-def print_time(func):
-     """
-     @print_time
-
-     Put this decorator around a function to see how many seconds each
-     call of this function takes to run.
-     """
-     def wrapped_func(*args, **kwargs):
-         start = time.time()
-         result = func(*args, **kwargs)
-         end = time.time()
-         seconds = end - start
-         print "SECONDS:", seconds, func.__name__, kwargs
-         return result
-     return wrapped_func
